+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/validation/GameValidationTexts.java
@@ -117,7 +117,7 @@ public class GameValidationTexts {
       }
       case CODE_ALT_SOUND_NOT_ENABLED: {
         label = "ALT sound not enabled.";
-        text = "The table has an ALT sound package, but it's not enabled in VPin MAME.";
+        text = "The table has an ALT sound package, but \"Sound Mode\" is not checked in the \"VPin MAME\" settings.";
         break;
       }
       case CODE_ALT_SOUND_FILE_MISSING: {